# Conflicts:
#	baystation12.dme
#	code/game/jobs/job/civilian.dm
#	code/game/jobs/job_controller.dm
#	code/game/objects/random/random.dm
#	code/game/objects/structures/lattice.dm
#	code/modules/client/preference_setup/occupation/occupation.dm
#	code/modules/power/cell.dm
#	code/modules/projectiles/guns/energy/laser.dm
#	code/modules/projectiles/guns/projectile/automatic.dm
#	icons/obj/gun.dmi
#	maps/torch/torch.dm

+++ code/datums/mind.dm
@@ -164,13 +164,26 @@
 
 /datum/mind/Topic(href, href_list)
 
-	if(check_rights(R_ADMIN) || usr == current)
+	if(href_list["add_goal"])
 
-		if(href_list["add_goal"])
-			log_admin("[key_name_admin(usr)] added a random goal to [key_name(current)].")
+		var/mob/caller = locate(href_list["add_goal_caller"])
+
+		var/is_admin =   FALSE
+		var/can_modify = FALSE
+		if(!caller || caller != current)
+			is_admin = check_rights(R_ADMIN)
+			can_modify = is_admin
+		else
+			can_modify = TRUE
+
+		if(can_modify)
+			if(is_admin)
+				log_admin("[key_name_admin(usr)] added a random goal to [key_name(current)].")
 			to_chat(current, SPAN_NOTICE("You have received a new goal. Use <b>Show Goals</b> to view it."))
 			generate_goals(assigned_job, TRUE, 1)
-			. = TRUE
+			return TRUE // To avoid 'you are not an admin' spam.
+
+	if(check_rights(R_ADMIN))
 
 		if(href_list["abandon_goal"])
 			var/datum/goal/goal = get_goal_from_href(href_list["abandon_goal"])
@@ -185,7 +198,7 @@
 
 		if(href_list["reroll_goal"])
 			var/datum/goal/goal = get_goal_from_href(href_list["reroll_goal"])
-			if(goal)
+			if(goal && (goal in goals))
 				qdel(goal)
 				generate_goals(assigned_job, TRUE, 1)
 				goal = goals[LAZYLEN(goals)]
@@ -229,9 +242,9 @@
 		if(antag) antag.place_mob(src.current)
 
 	else if (href_list["role_edit"])
-		var/new_role = input("Select new role", "Assigned role", assigned_role) as null|anything in job_master.occupations_by_title
+		var/new_role = input("Select new role", "Assigned role", assigned_role) as null|anything in SSjobs.titles_to_datums
 		if (!new_role) return
-		var/datum/job/job = job_master.occupations_by_title[new_role]
+		var/datum/job/job = SSjobs.get_by_title(new_role)
 		if(job)
 			assigned_job = job
 			assigned_role = job.title
@@ -562,7 +575,8 @@
 //HUMAN
 /mob/living/carbon/human/mind_initialize()
 	..()
-	if(!mind.assigned_role)	mind.assigned_role = "Assistant"	//defualt
+	if(!mind.assigned_role)
+		mind.assigned_role = GLOB.using_map.default_assistant_title
 
 //slime
 /mob/living/carbon/slime/mind_initialize()
